move state and commands out of main sphere file
